==> project_corona_real2/P21_Naver_Search.py <==
# -*- coding:utf-8 -*-

# 이제 앞에 찾은 단어들로 네이버 검색을 통해 기사 찾기
# https://www.hankyung.com/society/article/2020081519317 - 광화문 집회
# https://www.seoul.co.kr/news/newsView.php?id=20200831500128&wlog_tag3=naver - 사랑제일교회
